Create BMO Trait
Move BMO code to separate trait
