+++ sccpManTraits/bmoFunctions.php
@@ -1,0 +1,110 @@
+<?php
+
+namespace FreePBX\modules\Sccp_manager\sccpManTraits;
+
+trait bmoFunctions {
+    /* unused but FPBX API requires it */
+
+    public function doConfigPageInit($page) {
+        $this->doGeneralPost();
+    }
+
+    /* unused but FPBX API requires it */
+
+    public function install() {
+
+    }
+
+    /* unused but FPBX API requires it */
+
+    public function uninstall() {
+
+    }
+
+    /* unused but FPBX API requires it */
+
+    public function backup() {
+
+    }
+
+    /* unused but FPBX API requires it */
+
+    public function restore($backup) {
+
+    }
+
+    public function getActionBar($request) {
+        $buttons = array();
+        switch ($request['display']) {
+            case 'sccp_adv':
+                if (empty($request['tech_hardware'])) {
+                    break;
+                }
+                $buttons = array(
+                    'submit' => array(
+                        'name' => 'ajaxsubmit',
+                        'id' => 'ajaxsubmit',
+                        'value' => _("Save")
+                    ),
+                    'Save' => array(
+                        'name' => 'ajaxsubmit2',
+                        'id' => 'ajaxsubmit2',
+                        'stayonpage' => 'yes',
+                        'value' => _("Save + Continue")
+                    ),
+                    'cancel' => array(
+                        'name' => 'cancel',
+                        'id' => 'ajaxcancel',
+                        'data-search' => '?display=sccp_adv',
+                        'data-hash' => 'sccpdialplan',
+                        'value' => _("Cancel")
+                    ),
+                );
+                break;
+            case 'sccp_phone':
+                if (empty($request['tech_hardware'])) {
+                    break;
+                }
+                $buttons = array(
+                    'submit' => array(
+                        'name' => 'ajaxsubmit',
+                        'id' => 'ajaxsubmit',
+                        'value' => _("Save")
+                    ),
+                    'Save' => array(
+                        'name' => 'ajaxsubmit2',
+                        'id' => 'ajaxsubmit2',
+                        'stayonpage' => 'yes',
+                        'value' => _("Save + Continue")
+                    ),
+                    'cancel' => array(
+                        'name' => 'cancel',
+                        'id' => 'ajaxcancel',
+                        'data-search' => '?display=sccp_phone',
+                        'data-hash' => 'sccpdevice',
+                        'value' => _("Cancel")
+                    ),
+                );
+
+                break;
+            case 'sccpsettings':
+                $buttons = array(
+                    'submit' => array(
+                        'name' => 'ajaxsubmit',
+                        'id' => 'ajaxsubmit',
+                        'value' => _("Submit")
+                    ),
+                    'reset' => array(
+                        'name' => 'reset',
+                        'id' => 'ajaxcancel',
+                        'data-reload' => 'reload',
+                        'value' => _("Reset")
+                    ),
+                );
+
+                break;
+        }
+        return $buttons;
+    }
+}
+?>
